Implemented front-end access to back-end API
Implemented numerous fixes to various back-end API functions adding compatibility for bundles. 
retrieveMeta: New function made by combining the older function to fetch art and ratings for a game. Will now detect if a URL is a bundle and recursively check each game in the bundle to check for NSFW/SFW status. Art is successfully retrieved whether a URL to a game or bundle is supplied.

class Deals: Outward facing front-end API URL that is accessed via localhost/deals?num=... where num is the amount of games to return in the JSON response. The JSON response will contain a direct link to the cover art of the game by utilizing retrieveMeta and will also contain the determination ofthe rating for the game/bundle also using retrieveMeta. 

class Lowest: Outward facing front-end API URL that is accessed via localhost/lowest?games=game1%2Cgame2, etc. The game titles *must* be in the 'plains' format, which is a no-spaces, all lowercase version of the name of the game. Will return a JSON response with the supplied games and their lowest-ever historical prices as well as a direct link to the cover art and the rating for it using a similar mechanism as the Deals class.