Filter NSFW by default, skip bundles, etc.
Decided to filter out NSFW results by default, which leads to a more complicated logic to get deals. Deals will continue to be queried until the desired amount of SFW games (note: not bundles) is returned. Due to this new logic, loading times suffer somewhat, but I mitigated this by querying 10 games from the ITAD API at a time and then going off this list for logic before querying the next 10. Decided to skip bundles because of the significant load times incurred due to the recursive nature of searching each game for ratings.