Caching into DB and other optimizations
Games are now fetched when Flask is initially launched and then put on a schedule to refresh every hour with new games. On every refresh, games are cached into the SQLite database and any call to load the frontend page will now pull from the database for instantaneous page loading. I've set the amount of games fetched on each refresh to six by default, but it can be overridden by an explicit call with another value. This will be adjusted for genre-specific pages later on. Obsolete functions as well as obsolete parts of functions have been removed except for one that may be fixed by IsThereAnyDeals. SQL inserts or other user input is appropriately sanitized. At the moment, only the database for the webpage is implemented, but databases for game recommendations and user accounts will be implemented soon.