The creation of the recommendation system
In summary, the recommendation system works simply like this: each genre can have three states, dislike, neutral and like (-1, 0, 1). Disliked genres will be hidden from the user and liked genres will be suggested more often. These are called "affinities" to the genres. I created two functions so far, to modify the affinity values in the table and to retrieve them from the table. More details on their usage are in the comments in the functions. The front-end endpoint to access these functions is also explained in detail in the comments.